Renames

+++ packages/contracts/contracts/Dependencies/BaseMath.sol
@@ -2,5 +2,5 @@
 pragma solidity 0.8.17;
 
 contract BaseMath {
-    uint public constant DECIMAL_PRECISION = 1e18;
+    uint256 public constant DECIMAL_PRECISION = 1e18;
 }
