+++ packages/contracts/contracts/Interfaces/IEBTCToken.sol
@@ -11,7 +11,7 @@ interface IEBTCToken is IERC20, IERC2612 {
     event CdpManagerAddressChanged(address _cdpManagerAddress);
     event BorrowerOperationsAddressChanged(address _newBorrowerOperationsAddress);
 
-    event EBTCTokenBalanceUpdated(address _user, uint _amount);
+    event EBTCTokenBalanceUpdated(address _user, uint256 _amount);
 
     // --- Functions ---
